backend/requirements.txt
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.0
fastapi>=0.115.0
uvicorn>=0.32.0
httpx>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0
